#704 linking up the questionnaire and record tours with a simplier UI element
