+++ src/app/components/TourQuestionnaires/index.tsx
@@ -1,7 +1,12 @@
 import { TourPointer } from "components/TourPointer";
 import * as React from "react";
 import { useTranslation } from "react-i18next";
-import { TourStage } from "redux/modules/tour";
+import { TourStage, tourStageAction } from "redux/modules/tour";
+import { TourContainer } from "components/TourContainer";
+import { Button, Segment } from "semantic-ui-react";
+import RocketIcon from "./../../theme/rocket.inline.svg";
+import { useDispatch } from "react-redux";
+import ReactGA from "react-ga";
 
 interface IProps {
   id: string;
@@ -64,7 +69,49 @@ export const IntroduceQuestionnaireCreation = (p: {
         },
       ]}
       stage={TourStage.QUESTIONNAIRE_3}
-      transitionOnUnmount={null}
+      transitionOnUnmount={TourStage.QUESTIONNAIRE_4}
     />
+  );
+};
+
+export const WhatNextAfterQuestionnaireTour = (): JSX.Element => {
+  const { t } = useTranslation();
+  const dispatch = useDispatch();
+
+  const onClick = () => {
+    ReactGA.event({
+      category: "tour",
+      label: "record",
+      action: "started-whats-next",
+    });
+    dispatch(tourStageAction(TourStage.RECORD_1));
+  };
+
+  return (
+    <TourContainer stage={TourStage.QUESTIONNAIRE_4} transitionOnUnmount={null}>
+      <Segment
+        key="quickStart"
+        id="quick-start"
+        raised={true}
+        compact={true}
+        style={{ marginLeft: "auto", marginRight: "auto" }}
+      >
+        <h3 style={{ fontWeight: "normal" }}>
+          <RocketIcon style={{ width: "1rem", marginRight: ".3rem" }} />
+          {t("What next?")}
+        </h3>
+        <p>
+          {t("Nice one, the questionnaire has been added to your organisation")}
+        </p>
+        <p>
+          {t(
+            "Let's imitate a beneficiary and answer the questions, creating your first record"
+          )}
+        </p>
+        <Button primary={true} onClick={onClick}>
+          {t("New Record")}
+        </Button>
+      </Segment>
+    </TourContainer>
   );
 };
