+++ src/app/redux/modules/tour/index.ts
@@ -6,6 +6,7 @@ export enum TourStage {
   QUESTIONNAIRE_1 = 1,
   QUESTIONNAIRE_2,
   QUESTIONNAIRE_3,
+  QUESTIONNAIRE_4,
   RECORD_1,
   RECORD_2,
   RECORD_3,
